Tarefa1_2021li1g071

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

estaOrganizado :: [(Peca, Coordenadas)] -> Bool

A função estaOrganizado verifica se uma lista de [(Peca, Coordenadas)] se encontra organizada (onde caso seja falso retorna False)

Exemplos de utilização:

>>> estaOrganizado [(Bloco, (1,0)), (Bloco, (0,0))]
False
>>> estaOrganizado [(Bloco, (0,0)), (Bloco, (1,0))]
True

organizar :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função organizar vai organizar uma lista [(Peca, Coordenadas)] primeiramente pela coordenada y dos elementos e depois pela sua coordenada x

Exemplos de utilização:

>>> organizar [(Bloco, (0,0)), (Bloco, (0,2)), (Bloco, (1,0)), (Bloco, (1,2))]
[(X,(0,0)),(X,(1,0)),(X,(0,2)),(X,(1,2))]
>>> organizar [(Bloco, (3,0)), (Bloco, (0,1)), (Bloco, (0,0)), (Bloco, (1,2))]
[(X,(3,0)),(X,(0,0)),(X,(0,1)),(X,(1,2))]

Nota:

Mesmo a invocação anterior parecer dar um resultado falso, na função principal validaPotencialMapa é chamada sempre a função estaOrganizado para verificar se já se encontra organizada a lista

locatePiece :: Coordenadas -> [(Peca, Coordenadas)] -> Bool

A função locatePiece vai retornar True caso a lista tenha um par de coordenadas igual ao par de coordenadas recebido com a função

Exemplos de utilização:

>>> locatePiece (0,2) [(Bloco, (0,0)), (Bloco, (0,2)), (Bloco, (1,0)), (Bloco, (1,2))]
True
>>> locatePiece (1,0) [(Bloco, (0,0)), (Bloco, (2,0)), (Bloco, (0,1)), (Bloco, (1,1))]
False

returnPiece :: Coordenadas -> [(Peca, Coordenadas)] -> (Peca, Coordenadas)

A função returnPiece em junção com a locatePiece vai retornar o elemento da lista que terá um par de coordenadas igual ao par de coordenadas nos dado com a função

Exemplos de utilização:

>>> returnPiece (0,2) [(Bloco, (0,0)), (Bloco, (0,2)), (Bloco, (1,0)), (Bloco, (1,2))]
(Bloco, (0,2))
>>> returnPiece (1,0) [(Bloco, (0,0)), (Bloco, (2,0)), (Bloco, (0,1)), (Bloco, (1,1))]
(Vazio, (1,0))

biggestYAxis0 :: [(Peca, Coordenadas)] -> Int

A função biggestYAxis0 vai retornar o maior elemento da primeira coluna do mapa final, ou seja a maior coordenada y dos pares de coordenadas com coordenada x igual a 0

Exemplos de utilização:

>>> biggestYAxis0 [(Bloco, (0,0)), (Bloco, (0,2)), (Bloco, (1,0)), (Bloco, (1,2))]
2
>>> biggestYAxis0 [(Bloco, (0,0)), (Bloco, (2,0)), (Bloco, (0,1)), (Bloco, (1,1))]
1

biggestX :: [(Peca, Coordenadas)] -> Int

A função biggestX vai retornar o comprimento do mapa final, ou seja a maior coordenada x da lista

Exemplos de utilização:

>>> biggestX [(Bloco, (0,0)), (Bloco, (0,2)), (Bloco, (1,0)), (Bloco, (1,2))]
1
>>> biggestX [(Bloco, (0,0)), (Bloco, (2,0)), (Bloco, (0,1)), (Bloco, (1,1))]
2

coordenadaJaExistente :: Coordenadas -> [(Peca, Coordenadas)] -> Bool

A função coordenadaJaExistente vai ao receber um par de coordenadas e uma lista de [(Peca, Coordenadas)] verificar se existem nessa lista algum elemento com um par de coordenadas igual ao par de coordenadas dado à função

Exemplos de utilização:

>>> coordenadaJaExistente (0,0) [(Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
False
>>> coordenadaJaExistente (0,0) [(Bloco, (1,0)), (Bloco, (0,0)), (Bloco, (1,2))]
True

testePecasRepetidas :: [(Peca, Coordenadas)] -> Bool

Esta função testePecasRepetidas irá utilizar a função coordenadaJaExistente para verificar se existem multiplas declarações da mesma coordenada na lista de [(Peca, Coordenadas)] nos atribuída, retornando True caso existam multiplas declarações da(s) mesma(s) coordenada(s)

Exemplos de utilização:

>>> testePecasRepetidas [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
False
>>> testePecasRepetidas [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (1,2)), (Bloco, (1,2))]
True

existePorta2 :: [(Peca, Coordenadas)] -> Bool

A função existePorta2 vai ao receber uma lista de [(Peca, Coordenadas)] verificar se existe uma porta nessa lista

Exemplos de utilização:

>>> existePorta2 [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (1,2)), (Bloco, (1,2))]
False
>>> existePorta2 [(Porta, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
True

existePorta1 :: [(Peca, Coordenadas)] -> Bool

A função existePorta1 vai ao receber uma lista de [(Peca, Coordenadas)] em combinação com a função existePorta2 verificar se existe uma e só uma porta nessa lista

Exemplos de utilização:

>>> existePorta1 [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (1,2)), (Bloco, (1,2))]
False
>>> existePorta1 [(Porta, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
True
>>> existePorta1 [(Porta, (0,0)), (Porta, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
False

checkUnder :: Coordenadas -> [(Peca, Coordenadas)] -> Bool

A função checkUnder ao receber um par de coordenadas e uma lista de [(Peca, Coordenadas)] vai verificar dois pontos:

Se existe nessa lista o par de coordenadas logo em baixo do par recebido pela função

Caso o ponto acima seja verdade, se a Peca que se encontra nesse lugar não seja ou Vazio ou Porta

Caso passe em ambos os pontos irá retornar True, senão False

Exemplos de utilização:

>>> checkUnder (0,0) [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
False
>>> checkUnder (0,0) [(Caixa, (0,0)), (Bloco, (1,0)), (Porta, (0,1)), (Bloco, (1,2))]
False
>>> checkUnder (0,0) [(Caixa, (0,0)), (Porta, (1,0)), (Bloco, (0,1)), (Bloco, (1,2))]
True

caixaValida :: [(Peca, Coordenadas)] -> Bool

A função caixaValida irá em combinação com a função checkUnder ver se, caso a lista recebida contenha uma caixa, essa caixa é válida. Uma caixa é válida se:

A caixa se encontra em cima de uma outra caixa

A caixa se encontra em cima de um bloco

Exemplos de utilização:

>>> caixaValida [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
True
>>> caixaValida [(Caixa, (0,0)), (Bloco, (1,0)), (Bloco, (0,2)), (Bloco, (1,2))]
False
>>> caixaValida [(Caixa, (0,0)), (Porta, (1,0)), (Bloco, (0,1)), (Bloco, (1,2))]
True

mapaTemVaziosImp :: [(Peca, Coordenadas)] -> Bool

A função mapaTemVaziosImp vai ao receber uma lista de [(Peca, Coordenadas)] verificar se esta tem no mínimo uma peça vazio nele onde para isso vai testar por duas vertentes:

Caso a peça seja Vazio

Caso duas peças consecutivas na lista não sejam consecutivas no mapa final (por exemplo: [(Caixa, (0,0)), (Bloco, (2,0))] têm entre eles implicitamente uma peça Vazio)

Exemplos de utilização:

>>> mapaTemVaziosImp [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (0,1)), (Bloco, (1,1))]
False
>>> mapaTemVaziosImp [(Bloco, (0,0)), (Vazio, (1,0)), (Bloco, (0,1)), (Bloco, (1,1))]
True
>>> mapaTemVaziosImp [(Caixa, (0,0)), (Bloco, (2,0)), (Bloco, (0,1)), (Bloco, (1,1))]
True

checkUp :: Coordenadas -> [(Peca, Coordenadas)] -> Int -> Bool

A função checkUp vai verificar de forma similar à checkRight se existe uma base de blocos do mapa, porém não vai testar a peça de cima para não criar ciclos infinitos em algumas ocasiões

Funções usadas:

returnPiece

checkRight

checkUp

checkDown :: Coordenadas -> [(Peca, Coordenadas)] -> Int -> Bool

A função checkDown vai verificar de forma similar à checkRight se existe uma base de blocos do mapa, porém não vai testar a peça de cima para não criar ciclos infinitos em algumas ocasiões

Funções usadas:

returnPiece

checkRight

checkDown

checkRight :: Coordenadas -> [(Peca, Coordenadas)] -> Int -> Bool

A função checkRight vai verificar em três fase distintas se existe uma base de blocos do mapa:

Caso a peça à direita da que estamos a verificar for bloco, iremos chamar recursivamente a checkRight a esse bloco

Se a peça não for um bloco, vamos ver se a peça a baixo da que estamos é bloco, onde se ela o for utilizamos a função checkDown nessa peça

Se ambas as peças à direita e em baixo não for bloco, vamos verificar a que está em cima, onde caso ela seja bloco chamamos a função checkUp a essa peça

Para realizar esta verificação de peças nós utilizamos a função returnPiece dando-lhe o par de coordenadas e a lista que queremos verificar

Notas:

Com este método de verificação caso o mapa seja de tal forma construido que se cria um ciclo infinito nesta função, nas três funções de check (checkRight, checkDown e checkUp) se a peça que estamos a verificar se encontra na borda direita do mapa, retornamos True se a peça for um bloco e False se o contrário

baseBlocos :: [(Peca, Coordenadas)] -> Bool

A função baseBlocos vai verificar se existe uma base constituída apenas por blocos na lista recebida, onde para isso vai verificar primeiramente se a peça que se encontra na coordenada y maior da primeira coluna é um bloco, onde caso o seja irá posteriormente ver a peça à sua direita

Funções auxiliares utilizadas:

returnPiece para ver a peça que se encontra na coordenada pretendida

biggestYAxis0 que irá devolver a coordenada y do elemento mais abaixo da primeira coluna

biggestX que irá devolver o comprimento do mapa resultante da lista que nos é dada

checkRight para verificar a peça à direita da primeira peça da base de blocos

Exemplos de utilização:

>>> baseBlocos [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (0,1)), (Bloco, (1,1))]
True
>>> baseBlocos [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (1,1)), (Bloco, (2,1)), (Bloco, (3,1)), (Bloco, (3,0)), (Bloco, (4,0))]
True
>>> baseBlocos [(Bloco, (0,0)), (Bloco, (1,0)), (Bloco, (1,1)), (Bloco, (3,1)), (Bloco, (3,0)), (Bloco, (4,0))]
False

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

Função que a Tarefa1 se encontra a implementar. A validaPotencialMapa vai verificar se uma lista de [(Peca, Coordenadas)] consegue passar por 5 pontos distintos:

Testar se o mapa dispõe de peças repetidas, tal que irá utilizar a função testePecasRepetidas

Verificar se existe uma, e só uma porta na lista de [(Peca, Coordenadas)], onde será utilizada a função existePorta1 para o suposto

Ver se caso existirem caixas na lista de [(Peca, Coordenadas)], essas mesmas são válidas, utilizando a função caixaValida para o efeito

Verificar se existem (no mínimo um) espaços vazios na lista de [(Peca, Coordenadas)], sendo utilizada a função mapaTemVaziosImp para o teste

Testar se a base do mapa é constituído apenas por blocos e se é uma linha conínua sem falácias, onde será utilizada a função baseBlocos neste teste

Notas:

Para uma lista de [(Peca, Coordenadas)] ser válida terá que passar por todos os testes, onde caso falhe um deles a função retornará False e não testará os restantes casos

A função testePecasRepetidas é a única das 5 que se o caso de teste der falso irá retornar True